+++ history.md
@@ -2,9 +2,9 @@
 
 ## 3.0.23 (*)
 - feat: 更新 `yyl-base-webpack-config@0.2.20`
-* TODO: 如遇到 zepto 禁止构建并提示更换
-* TODO: 通过 webpack.config.js 添加 vconsole 无效
-* TODO: webpack 运行崩溃问题
+- feat: 如遇到 zepto 禁止构建并提示更换
+- TODO: 通过 webpack.config.js 添加 vconsole 无效
+- TODO: webpack 运行崩溃问题
 
 ## 3.0.22 (2021-06-10)